Move Twiss plotting into Twiss package
